requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
PyQt5>=5.15.0
torch>=2.0.0
torchvision>=0.15.0
scipy>=1.10.0
Pillow>=10.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0

